Added cover to bombsite A. The pit has been added to bombsite A. added a side part to CT tunnel to B site to give attackers a bit of cover. Made it easier to access the ladder onto bombsite A.
Added a third entrance to bombsite B.
changed cover on bombsite B to be more usable.
Extened the wall from the ladder to block sight lines.
updated radar.

fixed lighting, mesh, and cubemaps.
Moved camera pos to be in the map.
moved lighting props higher around the map to reduce them getting in the way of the player.